fix: Move function_starts_with to backtrace_support
