+++ src/utils.rs
@@ -236,53 +236,3 @@ pub fn debug_images() -> Vec<DebugImage> {
 
 #[cfg(feature = "with_backtrace")]
 pub use backtrace_support::{current_stacktrace, trim_stacktrace};
-
-/// Checks whether the function name starts with the given pattern.
-///
-/// In trait implementations, the original type name is wrapped in "_< ... >" and colons are
-/// replaced with dots. This function accounts for differences while checking.
-pub fn function_starts_with(mut func_name: &str, pattern: &str) -> bool {
-    if func_name.starts_with("_<") {
-        func_name = &func_name[2..];
-    }
-
-    if !func_name.is_char_boundary(pattern.len()) {
-        return false;
-    }
-
-    func_name
-        .chars()
-        .zip(pattern.chars())
-        .all(|(f, p)| f == p || f == '.' && p == ':')
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_function_starts_with() {
-        assert!(function_starts_with(
-            "futures::task_impl::std::set",
-            "futures::"
-        ));
-
-        assert!(!function_starts_with(
-            "futures::task_impl::std::set",
-            "tokio::"
-        ));
-    }
-
-    #[test]
-    fn test_function_starts_with_impl() {
-        assert!(function_starts_with(
-            "_<futures..task_impl..Spawn<T>>::enter::_{{closure}}",
-            "futures::"
-        ));
-
-        assert!(!function_starts_with(
-            "_<futures..task_impl..Spawn<T>>::enter::_{{closure}}",
-            "tokio::"
-        ));
-    }
-}
